Users can set their time zone when signing up.
Doing so will affect adding and listing reminders. When adding reminders,
the date string is parsed into the users time zone and when listing
reminders they will be displayed in the users time zone. This also means
that the reminders will be sent out at the right time in users time zone.

The migration that adds the time zone field to the users table will set the
time zone to Stockholm for all existing users. This is because all
registered users are currently in that time zone.