Don't pass nil to Time.zone.local_to_utc.
Doing so will fool simple_form in some way I don't understand with the
effect that the due date field is filled with the offset between the users
time zone and UTC in seconds.

I was able to reproduce this in a feature spec but not in a unit spec. It
sure feels like some magic is going on between `ActiveRecord` and
`simple_form`.

Closes #10.