Don't load all reminders after validation errors.
When a reminder would fail to save due to validation errors, all the user's
reminders would be rendered in a scrambled order in the list in the
resulting index view. The reason for this was setting the `@reminders`
instance variable to a `reload`ed version of the bare `reminders`
association on the `current_user` instead of
`reminders.uncancelled.ordered`.

Closes #9.