Persist due dates as they're entered by users.
After a due date string, entered by a user, was parsed into a due at time it
was previously thrown away. The effect of this was that when the edit form
was rendered, the when field would be filled with the UTC database datetime
string for the due_at column. If the form was submitted without being
changed the UTC date would then be parsed in the users time zone effectively
changing the due date of the reminder.

By persisting the due date as it was entered by the user, we can show the
same string again and let the user edit that instead. This should be more
straight-forward and less surprising.

* Replace FutureReminder decorator with the ReminderForm form object.

Closes #15