Allow users to create reminders via email.
Allow users to create reminders by sending email to DUEDATE@rmnd.io where
DUEDATE is a string that can be parsed by the time parser. The reminder
title will be set to whatever is in the subject line.

If the email comes from an address that doesn't belong to a registered user
or if the due date or title is invalid the email will be dropped and no
reminder will be created.

Creating repeating reminders via email is currently not possible. Creating
reminders by CCing or BCCing DUEDATE@rmnd.io is currently not possible.

When a reminder is successfully created, a confirmation email is sent out.
When a reminder can't be created, an error email is sent out.